// FrontEnd/src/app/services/models/bakeryfront-end-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Address } from '../models/address';
import { Category } from '../models/category';
import { ContactInfo } from '../models/contact-info';
export interface BakeryfrontEndResponse {
  Back_img?: string;
  Logo?: string;
  address?: Address;
  back_img?: string;
  categorys?: Array<Category>;
  desc?: string;
  id?: number;
  info?: ContactInfo;
  logo?: string;
  name?: string;
  owner?: number;
}