// FrontEnd/src/app/services/models/new-address.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface NewAddress {
  city: string;
  country: string;
  id?: number;
  name: string;
  number: string;
  postalCode: string;
  provincia: string;
  state: string;
  street: string;
  telNumber: string;
}