// FrontEnd/src/app/services/models/week.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { WeekDay } from '../models/week-day';
export interface Week {
  dom?: WeekDay;
  gio?: WeekDay;
  id?: number;
  lun?: WeekDay;
  mar?: WeekDay;
  mer?: WeekDay;
  sab?: WeekDay;
  ven?: WeekDay;
}