// FrontEnd/src/app/services/models/week-day.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LocalTime } from '../models/local-time';
export interface WeekDay {
  allDayOpen?: boolean;
  closed?: boolean;
  dayOfWeek?: string;
  eveningClosing?: LocalTime;
  eveningOpening?: LocalTime;
  id?: number;
  morningClosing?: LocalTime;
  morningOpening?: LocalTime;
}